Add files via upload
##This program does the following:
Creates a vector of length 5 filled with arbitrary integers from 0 to 10 using NumPy.
Creates a vector with values ​​ranging from 15 to 55 and print all values​​ except the first and last.
Creates a random array with 1000 elements and computes the average, variance, standard deviation of the array elements. 
Calculates the cumulative sum of the elements along a given axis, sum over rows for each of the 3 columns and sum over columns for each of the 2 rows of a given 3x3 array. 
Computes the multiplication of two given matrixes.
IDE used.... Jupyter Notebook.